requirements.txt
python-docx
pydantic
typer